[TESTS] JDR-Generator : Apprentissage développement SCALA
- UX WEB [port 5173]
- API MIDDLEWARE [port 8080]
- DB MYSQL [port 3306]
- API GEMINI [port 3000]
